feat: integrate Lexical rich text editor and update chat components
- Added support for rich text messages in the chat interface using Lexical.
- Updated ChatInput and related components to handle EditorState and rich text formats.
- Introduced AIRichTextMessage component for rendering rich text messages.
- Removed legacy LexicalRichEditor component and adjusted database schema to drop richTextSchema.
- Updated dependencies to include Lexical packages at version 0.33.1.
