+++ apps/desktop/layer/renderer/src/modules/ai/chat/components/displays/AIDisplayEntriesPart.tsx
@@ -343,7 +343,7 @@ export const AIDisplayEntriesPart = ({ part }: { part: AIDisplayEntriesTool }) =
 
   return (
     <Card className="mb-2 w-full min-w-0">
-      <div className="w-[9999px]" />
+      <div className="w-[9999px] max-w-[calc(var(--ai-chat-layout-width,65ch)_-120px)]" />
       <CardHeader>
         <CardTitle className="text-text flex items-center gap-2 text-xl font-semibold">
           <span className="text-lg">📰</span>
