+++ packages/internal/components/package.json
@@ -28,6 +28,13 @@
     "@follow/types": "workspace:*",
     "@follow/utils": "workspace:*",
     "@headlessui/react": "2.2.4",
+    "@lexical/code": "0.33.1",
+    "@lexical/link": "0.33.1",
+    "@lexical/list": "0.33.1",
+    "@lexical/mark": "0.33.1",
+    "@lexical/markdown": "0.33.1",
+    "@lexical/react": "0.33.1",
+    "@lexical/rich-text": "0.33.1",
     "@microflash/remark-callout-directives": "4.4.0",
     "@radix-ui/react-accordion": "1.2.11",
     "@radix-ui/react-avatar": "1.1.10",
@@ -60,6 +67,7 @@
     "input-otp": "1.4.2",
     "jotai": "2.12.5",
     "katex": "0.16.22",
+    "lexical": "0.33.1",
     "masonic": "4.1.0",
     "motion": "12.23.0",
     "react-blurhash": "0.3.0",
