+++ packages/internal/components/src/ui/lexical-rich-editor/index.tsx
@@ -1,0 +1,4 @@
+export { LexicalRichEditor } from "./LexicalRichEditor"
+export { KeyboardPlugin } from "./plugins"
+export { defaultLexicalTheme } from "./theme"
+export type { LexicalRichEditorProps, LexicalRichEditorRef } from "./types"
