+++ packages/internal/components/src/ui/lexical-rich-editor/plugins/index.ts
@@ -1,0 +1,1 @@
+export { KeyboardPlugin } from "./KeyboardPlugin"
