+++ packages/internal/components/src/ui/lexical-rich-editor/types.ts
@@ -1,0 +1,25 @@
+import type { EditorState, LexicalEditor } from "lexical"
+
+export interface LexicalRichEditorRef {
+  getEditor: () => LexicalEditor
+  focus: () => void
+  clear: () => void
+  isEmpty: () => boolean
+}
+
+export interface LexicalRichEditorProps {
+  placeholder?: string
+  className?: string
+  onChange?: (editorState: EditorState, editor: LexicalEditor) => void
+  onKeyDown?: (event: KeyboardEvent) => boolean
+  autoFocus?: boolean
+  namespace?: string
+  theme?: any
+  enabledPlugins?: {
+    history?: boolean
+    markdown?: boolean
+    list?: boolean
+    link?: boolean
+    autoFocus?: boolean
+  }
+}
