+++ packages/internal/database/src/drizzle/0034_curly_darkstar.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `ai_chat_messages` DROP COLUMN `rich_text_schema`;
