+++ packages/internal/database/src/drizzle/meta/_journal.json
@@ -239,6 +239,13 @@
       "when": 1753890574250,
       "tag": "0033_shiny_sebastian_shaw",
       "breakpoints": true
+    },
+    {
+      "idx": 34,
+      "version": "6",
+      "when": 1753937164043,
+      "tag": "0034_curly_darkstar",
+      "breakpoints": true
     }
   ]
 }
