+++ packages/internal/database/src/drizzle/migrations.js
@@ -34,6 +34,7 @@ import m0030 from "./0030_common_gabe_jones.sql"
 import m0031 from "./0031_kind_ikaris.sql"
 import m0032 from "./0032_orange_prima.sql"
 import m0033 from "./0033_shiny_sebastian_shaw.sql"
+import m0034 from "./0034_curly_darkstar.sql"
 import journal from "./meta/_journal.json"
 
 export default {
@@ -73,5 +74,6 @@ export default {
     m0031,
     m0032,
     m0033,
+    m0034,
   },
 }
